+++ src/databaseInterface/Repositories/__init__.py
@@ -1,4 +1,4 @@
-from .songRepo import SongRepository
-from .listenRepo import ListenRepository
+from src.databaseInterface.Repositories.songRepo import SongRepository
+from src.databaseInterface.Repositories.listenRepo import ListenRepository
 
 __all__ = ["SongRepository", "ListenRepository"]