+++ src/providerInterface/song/providers/__init__.py
@@ -1,4 +1,4 @@
-from .youtube import YoutubeProvider
-from .providerInterface import ProviderInterface
+from src.providerInterface.song.providers.youtube import YoutubeProvider
+from src.providerInterface.song.providers.providerInterface import ProviderInterface
 
-from .registry import get_provider, list_providers, add_provider
+from src.providerInterface.song.providers.registry import get_provider, list_providers, add_provider